Moved roster to its own controller.
